+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 # Changelog
 
+## [1.5.2] - 2025-04-15
+
+### Added
+
+- PD_SDK_DEBUG env var.  Set it to true to enable debugging of Pipedream Connect
+  API requests.  Simple sanitization is performed to prevent sensitive field leakage
+  but use caution.
+
 ## [1.5.1] - 2025-04-15
 
 ### Added